backend/testy/requirements/prod.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile prod.in
#
aiofiles==22.1.0
    # via testrail-migrator
aiohttp==3.8.3
    # via testrail-migrator
aiosignal==1.3.1
    # via aiohttp
allure-uploader-v2 @ git+https://gitlab-pub.yadro.com/testy/allure-uploader-v2.git@release/2.0.9
    # via -r prod.in
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   django
async-timeout==4.0.3
    # via aiohttp
attrs==24.2.0
    # via aiohttp
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r prod.in
    #   django-celery-beat
celery-progress==0.1.3
    # via -r prod.in
certifi==2024.7.4
    # via requests
channels==4.1.0
    # via
    #   -r prod.in
    #   channels-redis
channels-redis==4.2.0
    # via -r prod.in
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cron-descriptor==1.4.3
    # via django-celery-beat
django==4.2.13
    # via
    #   -r prod.in
    #   channels
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-js-asset
    #   django-model-utils
    #   django-notifications-hq
    #   django-pgtrigger
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
    #   jsonfield
django-celery-beat==2.5.0
    # via -r prod.in
django-cors-headers==3.13.0
    # via -r prod.in
django-filter==22.1
    # via -r prod.in
django-js-asset==2.2.0
    # via django-mptt
django-model-utils==4.5.1
    # via django-notifications-hq
django-mptt @ git+https://gitlab-pub.yadro.com/testy/django-mptt.git
    # via -r prod.in
django-notifications-hq==1.8.3
    # via -r prod.in
django-pgtrigger==4.11.1
    # via -r prod.in
django-redis==5.2.0
    # via -r prod.in
django-simple-history==3.1.1
    # via -r prod.in
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.1
    # via
    #   -r prod.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.2.0
    # via -r prod.in
drf-yasg==1.21.4
    # via -r prod.in
et-xmlfile==2.0.0
    # via openpyxl
factory-boy==3.2.1
    # via testrail-migrator
faker==30.8.2
    # via factory-boy
filelock==3.16.1
    # via -r prod.in
flake8==7.1.1
    # via flake8-copyright-validator
flake8-copyright-validator==0.0.1
    # via -r prod.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gevent==24.2.1
    # via -r prod.in
greenlet==3.0.3
    # via
    #   -r prod.in
    #   gevent
gunicorn==22.0.0
    # via
    #   -r prod.in
    #   uvicorn-worker
h11==0.14.0
    # via uvicorn
humanize==4.4.0
    # via -r prod.in
idna==3.7
    # via
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via flake8-copyright-validator
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.4
    # via coreschema
jsonfield==3.1.0
    # via django-notifications-hq
kombu==5.4.0
    # via celery
markupsafe==2.1.5
    # via jinja2
mccabe==0.7.0
    # via flake8
msgpack==1.0.8
    # via channels-redis
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
openpyxl==3.1.1
    # via plugin-example
orjson==3.10.1
    # via -r prod.in
packaging==24.1
    # via
    #   drf-yasg
    #   gunicorn
pillow==9.5.0
    # via -r prod.in
pluggy==1.0.0
    # via -r prod.in
plugin-example @ git+https://gitlab-pub.yadro.com/testy/plugin-example.git@release/0.2.0
    # via -r prod.in
prompt-toolkit==3.0.47
    # via click-repl
propcache==0.2.0
    # via yarl
psycopg==3.1.19
    # via -r prod.in
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   python-crontab
python-dotenv==0.21.0
    # via -r prod.in
pytimeparse==1.1.8
    # via -r prod.in
pytz==2024.1
    # via
    #   celery
    #   django-notifications-hq
    #   drf-yasg
redis==4.6.0
    # via
    #   -r prod.in
    #   channels-redis
    #   django-redis
requests==2.28.1
    # via
    #   allure-uploader-v2
    #   coreapi
ruamel-yaml==0.18.6
    # via drf-yasg
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
swapper==1.4.0
    # via django-notifications-hq
testrail-migrator @ git+https://gitlab-pub.yadro.com/testy/testrail_migrator.git@release/0.3.0
    # via -r prod.in
timeago==1.0.16
    # via -r prod.in
tqdm==4.64.1
    # via testrail-migrator
typing-extensions==4.12.2
    # via
    #   faker
    #   psycopg
tzdata==2024.1
    # via django-celery-beat
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.19
    # via requests
uvicorn==0.30.1
    # via uvicorn-worker
uvicorn-worker==0.2.0
    # via -r prod.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
websockets==12.0
    # via -r prod.in
yarl==1.17.1
    # via aiohttp
zipp==3.20.0
    # via importlib-metadata
zope-event==5.0
    # via gevent
zope-interface==7.0.1
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
